+++ SendBirdSyncManager.framework/Headers/SendBirdSyncManager.h
@@ -22,7 +22,5 @@
 #import "SBSMMessageFilter.h"
 
 #import "SBSMConstants.h"
-#import "SBSMOperationQueue.h"
-#import "SBSMOperation.h"
 #import "SBSMSyncManagerOptions.h"